User Auth Connection [Login & Register] Basic Implementation
Within this commit: A basic user interface was created with @mui and the Formik package. This included a home, register and login page. This commit also focused on connecting the application to InterAct's backend and saving user entries to a Mongo Atlas Database. Validation and error cases was NOT yet implemented within this commit.