Basic Implementation of Context API & State Management
Within this commit: This commit focused on handling user login authentication to update the global state of the application depending whether or not the user is authenticated or not, e.g. show/hide navbar links, allow access to profile page, etc. The Context API was adopted as a suitable alternative to using larger state management libraries such as Redux or MobX. This commit ships with server bugs and unimplemented errors - beware.