Stepper Component Modularized & Temp Profile Bug Fix
Within this commit: The @mui 'Stepper' component was implemented and broken down into a smaller, resusable component. Also in this commit was a temporary bug fix to the way the profile page gets and sends the Bearer token in its' GET request - by local storage and not by state/reducer. This will be changed in an upcoming refactor.